+++ src/games/wow/wdb2file.h
@@ -54,7 +54,7 @@ public:
     return std::string(reinterpret_cast<char*>(stringTable + stringOffset));
   }
 
-  std::vector<std::string> get(unsigned int recordIndex, const core::TableStructure * structure) const;
+  std::vector<std::string> get(size_t recordIndex, const core::TableStructure * structure) const;
 
 private:
 };
