+++ src/games/wow/wdc1file.h
@@ -54,7 +54,7 @@ public:
 
   bool close();
 
-  std::vector<std::string> get(unsigned int recordIndex, const core::TableStructure * structure) const;
+  std::vector<std::string> get(size_t recordIndex, const core::TableStructure * structure) const;
 
 private:
   enum FIELD_COMPRESSION
@@ -105,7 +105,7 @@ private:
 
   void readWDBC1Header();
 
-  bool readFieldValue(unsigned int recordIndex, unsigned int fieldIndex, uint arrayIndex, uint arraySize, unsigned int & result) const;
+  bool readFieldValue(size_t recordIndex, unsigned int fieldIndex, uint arrayIndex, uint arraySize, unsigned int & result) const;
   uint32 readBitpackedValue(field_storage_info info, unsigned char * recordOffset) const;
 
   header m_header;
