+++ src/games/wow/wdc2file.cpp
@@ -500,7 +500,7 @@ bool WDC2File::close()
   return WDB5File::close();
 }
 
-std::vector<std::string> WDC2File::get(unsigned int recordIndex, const core::TableStructure * structure) const
+std::vector<std::string> WDC2File::get(size_t recordIndex, const core::TableStructure * structure) const
 {
   std::vector<std::string> result;
   unsigned char * recordOffset = m_recordOffsets[recordIndex];
@@ -614,7 +614,7 @@ WDC2File::~WDC2File()
   close();
 }
 
-bool WDC2File::readFieldValue(unsigned int recordIndex, unsigned int fieldIndex, uint arrayIndex, uint arraySize, unsigned int & result) const
+bool WDC2File::readFieldValue(size_t recordIndex, unsigned int fieldIndex, uint arrayIndex, uint arraySize, unsigned int & result) const
 {
   unsigned char * recordOffset = m_recordOffsets[recordIndex];
   field_storage_info info = m_fieldStorageInfo[fieldIndex];
